moved scenes into their own files. Add code comments.
